Create a month-to-month bibliography of selected units

Create a bibliography in a newsletter format from month-to-month From either a specified unit or a set of units (e.g., A University -> Faculties -> Departments -> Groups). The function has the ability to create HTML output designed for emails and/or webpages.

Usage

CristinMonthly(
  zotero,
  unit.key,
  style = "apa-single-spaced",
  locale = "nn-NO",
  start.date = Sys.Date(),
  end.date = NULL,
  use.filter = TRUE,
  filter = NULL,
  nvi = TRUE,
  user.cards = FALSE,
  use.identifiers = TRUE,
  zotero.check = FALSE,
  lang = "nn",
  post = FALSE,
  create.email = FALSE,
  create.web = FALSE,
  subject = NULL,
  header = NULL,
  footer = NULL,
  replace.style = TRUE,
  width = 700,
  silent = FALSE,
  log = list()
)

Arguments

zotero

What Zotero library to use

unit.key

What unit to search for

style

Citation style to use for appended bibliography and/or citations, Default: 'apa-single-spaced'

locale

Desired language format of bibliography, Default: 'nn-NO'

start.date

Results created from specified date (YYYY-MM), Default: Sys.Date()

end.date

Results created before specified date (YYYY-MM), Default: NULL

use.filter

Filter out specific items (otherwise supported by `CristinWrangler`), Default: TRUE

filter

Filter out specific item types, will default to item types usually associated with NVI, Default: NULL

nvi

Filter out Cristin items not 1/2 in NVI, Default: TRUE

user.cards

Get business cards of contributors (currently INN), Default: FALSE

use.identifiers

Use if ISBN/DOI identifiers if enabled, Default: TRUE

zotero.check

Check if Cristin items exists in Zotero library , Default: FALSE

lang

Set Cristin language (nn, nb, en), Default: 'nn'

post

Post new items to specified Zotero library, Default: FALSE

create.email

Create HTML output for email, Default: FALSE

create.web

Create HTML output for web, Default: FALSE

subject

Add a custom subject to email, Default: NULL

header

Add a custom header to email, Default: NULL

footer

Add a custom footer to email, Default: NULL

replace.style

Email clients (e.g., Outlook) is poor at rendering CSS, and the default setting is to remove hanging indent, Default: TRUE

width

Width of email in pixels, Default: 700

silent

c2z is noisy, tell it to be quiet, Default: FALSE

log

A list for storing log elements, Default: list()

Value

A list with bibliography and information for defined units

Details

Please see https://oeysan.github.io/c2z/

See also

Examples

# \donttest{
  # Dummy example
  newsletter <- CristinMonthly(
    Zotero(
      id = "4827927",
      api = "RqlAmlH5l1KPghfCseAq1sQ1"
    ),
    "209.0.0.0"
  )
#> Searching for collections 
#> Searching for all items in library 
# }