Check for supported export formats from Cristin to Zotero

Used to filter Cristin categories that are not supported

Usage

CristinSupported(
  data = NULL,
  zotero.check = TRUE,
  items = NULL,
  remove.na = TRUE,
  replace.na = "book",
  force.type = NULL,
  remove.duplicates = TRUE,
  silent = FALSE,
  log = list()
)

Arguments

data

Zotero-type matrix, Default: NULL

zotero.check

Should the function look for duplicates?, Default: TRUE

items

Items in library to check for

remove.na

Cristin contains many, more or less, obscure categories, and not all are (yet) supported. By default these are removed, however, if this option is set to FALSE unsupported categories are treated according to replace.na, Default: TRUE

replace.na

May the odds be in your favor and replace unsupported categories with a predefined itemType if remove.na is set to false, Default: 'book'

force.type

Force all items to a predefined itemType, Default: NULL

remove.duplicates

Remove duplicates if TRUE, Default: TRUE

silent

c2z is noisy, tell it to be quiet, Default: FALSE

log

A list for storing log elements, Default: list()

Value

Zotero supported items with unsupported categories as NA

Details

Please see https://oeysan.github.io/c2z/

See also

Examples

# Supported Cristin items
print(CristinSupported(), n = 5)
#> # A tibble: 33 × 2
#>   cristin.category zotero.types
#>   <chr>            <chr>       
#> 1 ANTHOLOGYACA     book        
#> 2 BOOKTRANSL       book        
#> 3 MONOGRAPHACA     book        
#> 4 COMMENTARYACA    book        
#> 5 TEXTBOOK         book        
#> # ℹ 28 more rows