Wrangle Cristin metadata into Zotero-type format

The little engine that could, at least try, to wrangle metadata from Cristin into a Zotero-type matrix. References can be augmented, or replaced with metadata gathered through ISBN or DOI identifiers, and through Crossref queries if all other fails. Or, hope the best and place your misguided faith in the ability of researchers to correctly register their own publications.

Usage

CristinWrangler(
  data,
  meta = list(),
  use.identifiers = TRUE,
  crossref.search = FALSE,
  autosearch = TRUE,
  override = FALSE,
  silent = FALSE,
  polite = TRUE,
  prefer.semantic = TRUE,
  log = list()
)

Arguments

data

Tibble containing metadata from Cristin

meta

A list for storing data elements, Default: list()

use.identifiers

Use if ISBN/DOI identifiers if enabled, Default: TRUE

crossref.search

Query Crossref database based on title, authors, and date if enabled, Default: FALSE

autosearch

Results could be automatically evaluated (based on some logic) or you could inspect them manually if set to FALSE, Default: TRUE

override

Put your faith in the algorithms and the identifiers (i.e., DOI/ISBN) and override what is reported in Cristin, Default: FALSE

silent

Running silent, running deep, Default: FALSE

polite

Please store you email in `.Renviron` to query Crossref, Default: TRUE

prefer.semantic

Prefer metadata from Semantic Scholar, Default: TRUE

log

A list for storing log elements, Default: list()

Value

A Zotero-type matrix (tibble)

Details

Please see https://oeysan.github.io/c2z/

See also

Examples

# \donttest{
  # Simple `Cristin` search by id with import set as FALSE
  example <- Cristin(id = "840998", zotero.import = FALSE)
#> Found 1 result 
#> Checking whether references are supported. See `CristinSupported()` 
#> Looking for missing data 

  if (any(nrow(example$results))) {
    # Use `ZoteroIndex` to print `CristinWrangler`
    CristinWrangler(example$results)$results |>
      ZoteroIndex() |>
      dplyr::select(name) |>
      print(width = 80)
  }
#> Converting 1 reference from Cristin to Zotero format 
#> —————————————————Process: 100.00% (1/1). Elapsed time: 00:00:01—————————————————
#> # A tibble: 1 × 1
#>   name                                                                          
#>   <chr>                                                                         
#> 1 Petersen (2012) Seeds of Satan: Conceptions of Magic in Contemporary Satanism…
# }