Wrangle CrossRef metadata into Zotero-type format

Query CrossRef by DOI and fetch metadata

Usage

DoiCrossref(data, meta = list(), use.json = TRUE, silent = TRUE, log = list())

Arguments

data

XML data from CrossRef containing metadata

meta

A list collecting all metadata used to create , Default: list()

use.json

Use either json (TRUE) or XML (FALSE), Default: TRUE

silent

c2z is noisy, tell it to be quiet, Default: TRUE

log

A list for storing log elements, Default: list()

Value

A Zotero-type matrix (tibble)

Details

Please see https://oeysan.github.io/c2z/

Examples

# \donttest{
  # `DoiCrossref` is called from `ZoteroDoi` (if item found)
  example <- ZoteroDoi("10.1126/sciadv.abd1705")

  # Print index using `ZoteroIndex`
  if (!is.null(example$data)) {
    ZoteroIndex(example$data) |>
      dplyr::select(name) |>
      print(width = 80)

    # Display catalog
    example$data$libraryCatalog
  }
#> # A tibble: 1 × 1
#>   name                                                                          
#>   <chr>                                                                         
#> 1 Serra-Garcia & Gneezy (2021) Nonreplicable publications are cited more than r…
#> [1] "DOI.org (Crossref)"
# }