Add items to Zotero list

Use identifiers or predefined data to add to the Zotero list

Usage

ZoteroAdd(zotero, metadata = NULL, doi = NULL, isbn = NULL, silent = FALSE)

Arguments

zotero

A list with information on the specified Zotero library (e.g., id, API key, collections, and items)

metadata

Predefined metadata in Zoter-format, Default: NULL

doi

Use ZoteroDoi to fetch DOI metadata, Default: NULL

isbn

Use ZoteroIsbn to fetch ISBN metadata, Default: NULL

silent

c2z is noisy, tell it to be quiet, Default: FALSE

Value

Will add data to the Zotero list

Details

Please see https://oeysan.github.io/c2z/

Examples

# \donttest{
  # Add items from `ZoteroIsbn` to the default Zotero list
  example <- ZoteroAdd(
    Zotero(
      id = "9913421",
      api = "RqlAmlH5l1KPghfCseAq1sQ1"
    ),
    isbn = "978-1529797138"
  )
#> Searching 1 item using ISBN 
  # Print index using `ZoteroIndex`
  if (any(nrow(example$items))) {
    ZoteroIndex(example$items) |>
      dplyr::select(name) |>
      print(width = 80)
  }
#> # A tibble: 1 × 1
#>   name                                                        
#>   <chr>                                                       
#> 1 Field (2022) Adventure in statistics: the reality enigma, An
# }