Check Zotero library for duplicates

Remove references that are not modified since last added to Zotero

Usage

ZoteroCheck(
  data,
  id,
  id.type,
  created,
  last.modified,
  items,
  remove.duplicates = TRUE,
  silent = FALSE,
  log = list()
)

Arguments

data

Tibble containing metadata (from Cristin)

id

column containing identifier (e.g., cristin_result_id)

id.type

Type of metadata as found in the Zotero extra field ("e.g., Cristin)

created

column containing creation date in UNIX timestamp format (e.g., created)

last.modified

column containing modification date in UNIX timestamp format (e.g., last_modified)

items

Items in library to check for

remove.duplicates

Remove duplicates if TRUE, Default: TRUE

silent

c2z is noisy, tell it to be quiet, Default: FALSE

log

A list for storing log elements, Default: list()

Value

Returns non-duplicated data in a Zotero-type matrix (tibble)

Details

Please see https://oeysan.github.io/c2z/

See also

Examples

# \donttest{
  # Simple `Cristin` search by id
  cristin.data <- Cristin(
    id = "840998",
    zotero.import = FALSE
  )
#> Found 1 result 
#> Checking whether references are supported. See `CristinSupported()` 
#> Looking for missing data 

  # Simple `ZoteroCheck`
  example <- ZoteroCheck(
    data = cristin.data$result,
    id = "cristin_result_id",
    id.type = "Cristin",
    created = "created",
    last.modified = "last_modified",
    items = Zotero(
      user = FALSE,
      id = "4827927",
      api = "RqlAmlH5l1KPghfCseAq1sQ1",
      library = TRUE,
      silent = TRUE
    )$items
  )
#> Checking whether references exist in library 
# }