Copy collections and items from a Zotero library

Replace key identifiers with new ones while keeping existing structure and relationship between collections and items

Usage

ZoteroCopy(
  zotero,
  copy.collections = TRUE,
  copy.items = TRUE,
  copy.extras = TRUE,
  remove.missing = TRUE,
  change.library = FALSE,
  copy.user = TRUE,
  copy.id = NULL,
  copy.api = NULL,
  silent = FALSE
)

Arguments

zotero

A list with information on the specified Zotero library (e.g., id, API key, collections, and items)

copy.collections

Try to copy specified collections, Default: TRUE

copy.items

Try to copy specified items?, Default: TRUE

copy.extras

Try to copy specified extras (i.e., attachments and notes)?, Default: TRUE

remove.missing

Deleted missing extras, Default: TRUE

change.library

Stage changing of library (e.g., from a group to a personal library), Default: FALSE

copy.user

New user type (The functions will use `group` as prefix if FALSE), Default: TRUE

copy.id

New id, Default: NULL

copy.api

New API key, Default: NULL

silent

c2z is noisy, tell it to be quiet, Default: FALSE

Value

A list with information on the specified Zotero library (e.g., copied collections and items)

Details

Please see https://oeysan.github.io/c2z/

See also

Examples

# \donttest{
  # Fetching collections and items from default group
  zotero = Zotero(
    user = FALSE,
    id = "4827927",
    api = "RqlAmlH5l1KPghfCseAq1sQ1",
    library = TRUE
  )
#> Searching for collections 
#> Found 2 collections 
#> The Zotero list contains: 2 collections, 0 items, and 0 attachments 
#> Searching for all items in library 
#> Found 4 items 
#> The Zotero list contains: 2 collections, 2 items, and 2 attachments 

  # Display collections
  if (any(nrow(zotero$collections))) {
    zotero$collections |>
      dplyr::select(key, version, parentCollection) |>
      print(width = 80)
  }
#> # A tibble: 2 × 3
#>   key      version parentCollection
#>   <chr>      <int> <chr>           
#> 1 9QJNRU53      11 FALSE           
#> 2 4UKXTQJ4      12 9QJNRU53        

  # Display items
  if (any(nrow(zotero$items))) {
    zotero$items |>
      dplyr::select(key, version) |>
      print(width = 80)
  }
#> # A tibble: 4 × 2
#>   key      version
#>   <chr>      <int>
#> 1 WH8UMVIP      29
#> 2 JQXF6FH7      38
#> 3 K2PA6HTE      20
#> 4 IJN372W3      18

  # Copy items
  example <- ZoteroCopy(
    zotero,
  )
#> Copying collections 
#> Copying items 

  # Display collections
  if (any(nrow(example$collections))) {
    example$collections |>
      dplyr::select(key, version, parentCollection) |>
      print(width = 80)
  }
#> # A tibble: 2 × 3
#>   key      version parentCollection
#>   <chr>      <dbl> <chr>           
#> 1 53NF3AZZ       0 FALSE           
#> 2 KR5HH76M       0 53NF3AZZ        

  # Display items
  if (any(nrow(example$items))) {
    example$items |>
      dplyr::select(key, version) |>
      print(width = 80)
  }
#> # A tibble: 4 × 2
#>   key      version
#>   <chr>      <dbl>
#> 1 Y2KJESUZ       0
#> 2 WPZXZ8CX       0
#> 3 99NF4LJS       0
#> 4 VJ8399UQ       0
# }