Export Zotero items to bibliography

Export Zotero items to a specified format (e.g., BibLaTeX) using a CSL format (e.g., APA7)

Usage

ZoteroExport(
  zotero,
  csl.type = NULL,
  csl.name = "style",
  locale = "en-US",
  format = "biblatex",
  bib.name = "references",
  save.data = FALSE,
  save.path = NULL,
  silent = FALSE
)

Arguments

zotero

A list with information on the specified Zotero library (e.g., id, API key, collections, and items)

csl.type

Specify a CSL type to Official repository for Citation Style Language (CSL), Default: NULL

csl.name

Name of saved CSL file, Default: 'style'

locale

Desired language format of bibliography, Default: 'en-US'

format

Export format of Zotero items, Default: 'biblatex'

bib.name

Name of exported bibliography, Default: 'references'

save.data

Save data (e.g., bibliography) to disk, Default: FALSE

save.path

Location to store data on disk, Default: NULL

silent

c2z is noisy, tell it to be quiet, Default: FALSE

Value

A list with information on the specified Zotero library (e.g., exported items, bibliography and citations)

Details

Please see https://oeysan.github.io/c2z/

See also

Examples

# \donttest{
  # Define Zotero list according to default group
  zotero = Zotero(
    user = FALSE,
    id = "4827927",
    api = "RqlAmlH5l1KPghfCseAq1sQ1",
    item.type = "-attachment || note",
    max.result = 1,
    library = TRUE
  )
#> Searching for collections 
#> Found 11 collections 
#> The Zotero list contains: 11 collections, 0 items, and 0 attachments 
#> Searching for all items in library 
#> Found 9 items 
#> The provided query is limited to 1 result 
#> The Zotero list contains: 11 collections, 1 item, and 0 attachments 

  # Export 1 items from the default group
  example <- ZoteroExport(
    zotero
  )
#> Found 1 `biblatex` reference 
  # Display exported
  cat(example$export, fill = 80)
#> 
#> @article{noauthor_bibliometrix_2017,
#> 	title = {bibliometrix : An R-tool for comprehensive science mapping analysis},
#> 	volume = {11},
#> 	issn = {1751-1577},
#> 	url = {https://doi.org/10.1016/j.joi.2017.08.007},
#> 	doi = {https://doi.org/10.1016/j.joi.2017.08.007},
#> 	abstract = {The use of bibliometrics is gradually extending to all disciplines. It is particularly suitable for science mapping at a time when the emphasis on empirical contributions is producing voluminous, fragmented, and controversial research streams. Science mapping is complex and unwieldly because it is multi-step and frequently requires numerous and diverse software tools, which are not all necessarily freeware. Although automated workflows that integrate these software tools into an organized data flow are emerging, in this paper we propose a unique open-source tool, designed by the authors, called bibliometrix, for performing comprehensive science mapping analysis. bibliometrix supports a recommended workflow to perform bibliometric analyses. As it is programmed in R, the proposed tool is flexible and can be rapidly upgraded and integrated with other statistical R-packages. It is therefore useful in a constantly changing science such as bibliometrics.},
#> 	pages = {959--975},
#> 	number = {4},
#> 	journaltitle = {Journal of Informetrics},
#> 	date = {2017-11-01},
#> 	langid = {english},
#> 	note = {grants:{NA}
#> cited\_by\_count:3519
#> counts\_by\_year:2023:2016
#> referenced\_works:https://openalex.org/W767067438
#> related\_works:https://openalex.org/W45233828
#> is\_paratext:{FALSE}
#> is\_retracted:{FALSE}
#> grants:{NA}
#> cited\_by\_count:3519
#> counts\_by\_year:c(1215, 1131, 692, 294, 71, 20, 1, 1)
#> referenced\_works:https://openalex.org/W1799889357
#> related\_works:https://openalex.org/W144424372
#> is\_paratext:{FALSE}
#> is\_retracted:{FALSE}
#> grants:{NA}
#> cited\_by\_count:3519
#> counts\_by\_year:2023:2016
#> referenced\_works:https://openalex.org/W1963753783
#> related\_works:https://openalex.org/W188202134
#> is\_paratext:{FALSE}
#> is\_retracted:{FALSE}
#> grants:{NA}
#> cited\_by\_count:3519
#> counts\_by\_year:c(1215, 1131, 692, 294, 71, 20, 1, 1)
#> referenced\_works:https://openalex.org/W1966104031
#> related\_works:https://openalex.org/W1981780420
#> is\_paratext:{FALSE}
#> is\_retracted:{FALSE}
#> grants:{NA}
#> cited\_by\_count:3519
#> counts\_by\_year:2023:2016
#> referenced\_works:https://openalex.org/W1966538856
#> related\_works:https://openalex.org/W2105666325
#> is\_paratext:{FALSE}
#> is\_retracted:{FALSE}
#> grants:{NA}
#> cited\_by\_count:3519
#> counts\_by\_year:c(1215, 1131, 692, 294, 71, 20, 1, 1)
#> referenced\_works:https://openalex.org/W1968283201
#> related\_works:https://openalex.org/W2182707996
#> is\_paratext:{FALSE}
#> is\_retracted:{FALSE}
#> grants:{NA}
#> cited\_by\_count:3519
#> counts\_by\_year:2023:2016
#> referenced\_works:https://openalex.org/W1975107573
#> related\_works:https://openalex.org/W2885845152
#> is\_paratext:{FALSE}
#> is\_retracted:{FALSE}
#> grants:{NA}
#> cited\_by\_count:3519
#> counts\_by\_year:c(1215, 1131, 692, 294, 71, 20, 1, 1)
#> referenced\_works:https://openalex.org/W1976620775
#> related\_works:https://openalex.org/W2964988449
#> is\_paratext:{FALSE}
#> is\_retracted:{FALSE}
#> grants:{NA}
#> cited\_by\_count:3519
#> counts\_by\_year:2023:2016
#> referenced\_works:https://openalex.org/W1977714176
#> related\_works:https://openalex.org/W4240783740
#> is\_paratext:{FALSE}
#> is\_retracted:{FALSE}
#> grants:{NA}
#> cited\_by\_count:3519
#> counts\_by\_year:c(1215, 1131, 692, 294, 71, 20, 1, 1)
#> referenced\_works:https://openalex.org/W1981340255
#> related\_works:https://openalex.org/W4386031268
#> is\_paratext:{FALSE}
#> is\_retracted:{FALSE}
#> grants:{NA}
#> cited\_by\_count:3519
#> counts\_by\_year:2023:2016
#> referenced\_works:https://openalex.org/W1981886524
#> related\_works:https://openalex.org/W45233828
#> is\_paratext:{FALSE}
#> is\_retracted:{FALSE}
#> grants:{NA}
#> cited\_by\_count:3519
#> counts\_by\_year:c(1215, 1131, 692, 294, 71, 20, 1, 1)
#> referenced\_works:https://openalex.org/W1989669128
#> related\_works:https://openalex.org/W144424372
#> is\_paratext:{FALSE}
#> is\_retracted:{FALSE}
#> grants:{NA}
#> cited\_by\_count:3519
#> counts\_by\_year:2023:2016
#> referenced\_works:https://openalex.org/W1996229960
#> related\_works:https://openalex.org/W188202134
#> is\_paratext:{FALSE}
#> is\_retracted:{FALSE}
#> grants:{NA}
#> cited\_by\_count:3519
#> counts\_by\_year:c(1215, 1131, 692, 294, 71, 20, 1, 1)
#> referenced\_works:https://openalex.org/W2002117998
#> related\_works:https://openalex.org/W1981780420
#> is\_paratext:{FALSE}
#> is\_retracted:{FALSE}
#> grants:{NA}
#> cited\_by\_count:3519
#> counts\_by\_year:2023:2016
#> referenced\_works:https://openalex.org/W2017680781
#> related\_works:https://openalex.org/W2105666325
#> is\_paratext:{FALSE}
#> is\_retracted:{FALSE}
#> grants:{NA}
#> cited\_by\_count:3519
#> counts\_by\_year:c(1215, 1131, 692, 294, 71, 20, 1, 1)
#> referenced\_works:https://openalex.org/W2027998681
#> related\_works:https://openalex.org/W2182707996
#> is\_paratext:{FALSE}
#> is\_retracted:{FALSE}
#> grants:{NA}
#> cited\_by\_count:3519
#> counts\_by\_year:2023:2016
#> referenced\_works:https://openalex.org/W2034800299
#> related\_works:https://openalex.org/W2885845152
#> is\_paratext:{FALSE}
#> is\_retracted:{FALSE}
#> grants:{NA}
#> cited\_by\_count:3519
#> counts\_by\_year:c(1215, 1131, 692, 294, 71, 20, 1, 1)
#> referenced\_works:https://openalex.org/W2042852414
#> related\_works:https://openalex.org/W2964988449
#> is\_paratext:{FALSE}
#> is\_retracted:{FALSE}
#> grants:{NA}
#> cited\_by\_count:3519
#> counts\_by\_year:2023:2016
#> referenced\_works:https://openalex.org/W2045108252
#> related\_works:https://openalex.org/W4240783740
#> is\_paratext:{FALSE}
#> is\_retracted:{FALSE}
#> grants:{NA}
#> cited\_by\_count:3519
#> counts\_by\_year:c(1215, 1131, 692, 294, 71, 20, 1, 1)
#> referenced\_works:https://openalex.org/W2052710775
#> related\_works:https://openalex.org/W4386031268
#> is\_paratext:{FALSE}
#> is\_retracted:{FALSE}
#> grants:{NA}
#> cited\_by\_count:3519
#> counts\_by\_year:2023:2016
#> referenced\_works:https://openalex.org/W2056147580
#> related\_works:https://openalex.org/W45233828
#> is\_paratext:{FALSE}
#> is\_retracted:{FALSE}
#> grants:{NA}
#> cited\_by\_count:3519
#> counts\_by\_year:c(1215, 1131, 692, 294, 71, 20, 1, 1)
#> referenced\_works:https://openalex.org/W2066888021
#> related\_works:https://openalex.org/W144424372
#> is\_paratext:{FALSE}
#> is\_retracted:{FALSE}
#> grants:{NA}
#> cited\_by\_count:3519
#> counts\_by\_year:2023:2016
#> referenced\_works:https://openalex.org/W2069134922
#> related\_works:https://openalex.org/W188202134
#> is\_paratext:{FALSE}
#> is\_retracted:{FALSE}
#> grants:{NA}
#> cited\_by\_count:3519
#> counts\_by\_year:c(1215, 1131, 692, 294, 71, 20, 1, 1)
#> referenced\_works:https://openalex.org/W2075220720
#> related\_works:https://openalex.org/W1981780420
#> is\_paratext:{FALSE}
#> is\_retracted:{FALSE}
#> grants:{NA}
#> cited\_by\_count:3519
#> counts\_by\_year:2023:2016
#> referenced\_works:https://openalex.org/W2094864959
#> related\_works:https://openalex.org/W2105666325
#> is\_paratext:{FALSE}
#> is\_retracted:{FALSE}
#> grants:{NA}
#> cited\_by\_count:3519
#> counts\_by\_year:c(1215, 1131, 692, 294, 71, 20, 1, 1)
#> referenced\_works:https://openalex.org/W2097148950
#> related\_works:https://openalex.org/W2182707996
#> is\_paratext:{FALSE}
#> is\_retracted:{FALSE}
#> grants:{NA}
#> cited\_by\_count:3519
#> counts\_by\_year:2023:2016
#> referenced\_works:https://openalex.org/W2098162425
#> related\_works:https://openalex.org/W2885845152
#> is\_paratext:{FALSE}
#> is\_retracted:{FALSE}
#> grants:{NA}
#> cited\_by\_count:3519
#> counts\_by\_year:c(1215, 1131, 692, 294, 71, 20, 1, 1)
#> referenced\_works:https://openalex.org/W2108680868
#> related\_works:https://openalex.org/W2964988449
#> is\_paratext:{FALSE}
#> is\_retracted:{FALSE}
#> grants:{NA}
#> cited\_by\_count:3519
#> counts\_by\_year:2023:2016
#> referenced\_works:https://openalex.org/W2128438887
#> related\_works:https://openalex.org/W4240783740
#> is\_paratext:{FALSE}
#> is\_retracted:{FALSE}
#> grants:{NA}
#> cited\_by\_count:3519
#> counts\_by\_year:c(1215, 1131, 692, 294, 71, 20, 1, 1)
#> referenced\_works:https://openalex.org/W2135455887
#> related\_works:https://openalex.org/W4386031268
#> is\_paratext:{FALSE}
#> is\_retracted:{FALSE}
#> grants:{NA}
#> cited\_by\_count:3519
#> counts\_by\_year:2023:2016
#> referenced\_works:https://openalex.org/W2146695321
#> related\_works:https://openalex.org/W45233828
#> is\_paratext:{FALSE}
#> is\_retracted:{FALSE}
#> grants:{NA}
#> cited\_by\_count:3519
#> counts\_by\_year:c(1215, 1131, 692, 294, 71, 20, 1, 1)
#> referenced\_works:https://openalex.org/W2150220236
#> related\_works:https://openalex.org/W144424372
#> is\_paratext:{FALSE}
#> is\_retracted:{FALSE}
#> grants:{NA}
#> cited\_by\_count:3519
#> counts\_by\_year:2023:2016
#> referenced\_works:https://openalex.org/W2171458572
#> related\_works:https://openalex.org/W188202134
#> is\_paratext:{FALSE}
#> is\_retracted:{FALSE}
#> grants:{NA}
#> cited\_by\_count:3519
#> counts\_by\_year:c(1215, 1131, 692, 294, 71, 20, 1, 1)
#> referenced\_works:https://openalex.org/W2225280195
#> related\_works:https://openalex.org/W1981780420
#> is\_paratext:{FALSE}
#> is\_retracted:{FALSE}
#> grants:{NA}
#> cited\_by\_count:3519
#> counts\_by\_year:2023:2016
#> referenced\_works:https://openalex.org/W2285085441
#> related\_works:https://openalex.org/W2105666325
#> is\_paratext:{FALSE}
#> is\_retracted:{FALSE}
#> grants:{NA}
#> cited\_by\_count:3519
#> counts\_by\_year:c(1215, 1131, 692, 294, 71, 20, 1, 1)
#> referenced\_works:https://openalex.org/W2408216567
#> related\_works:https://openalex.org/W2182707996
#> is\_paratext:{FALSE}
#> is\_retracted:{FALSE}
#> grants:{NA}
#> cited\_by\_count:3519
#> counts\_by\_year:2023:2016
#> referenced\_works:https://openalex.org/W2472756135
#> related\_works:https://openalex.org/W2885845152
#> is\_paratext:{FALSE}
#> is\_retracted:{FALSE}
#> grants:{NA}
#> cited\_by\_count:3519
#> counts\_by\_year:c(1215, 1131, 692, 294, 71, 20, 1, 1)
#> referenced\_works:https://openalex.org/W2963453445
#> related\_works:https://openalex.org/W2964988449
#> is\_paratext:{FALSE}
#> is\_retracted:{FALSE}
#> grants:{NA}
#> cited\_by\_count:3519
#> counts\_by\_year:2023:2016
#> referenced\_works:https://openalex.org/W3014861942
#> related\_works:https://openalex.org/W4240783740
#> is\_paratext:{FALSE}
#> is\_retracted:{FALSE}
#> grants:{NA}
#> cited\_by\_count:3519
#> counts\_by\_year:c(1215, 1131, 692, 294, 71, 20, 1, 1)
#> referenced\_works:https://openalex.org/W3103443220
#> related\_works:https://openalex.org/W4386031268
#> is\_paratext:{FALSE}
#> is\_retracted:{FALSE}
#> grants:{NA}
#> cited\_by\_count:3519
#> counts\_by\_year:2023:2016
#> referenced\_works:https://openalex.org/W3125707221
#> related\_works:https://openalex.org/W45233828
#> is\_paratext:{FALSE}
#> is\_retracted:{FALSE}
#> grants:{NA}
#> cited\_by\_count:3519
#> counts\_by\_year:c(1215, 1131, 692, 294, 71, 20, 1, 1)
#> referenced\_works:https://openalex.org/W4241271818
#> related\_works:https://openalex.org/W144424372
#> is\_paratext:{FALSE}
#> is\_retracted:{FALSE}
#> grants:{NA}
#> cited\_by\_count:3519
#> counts\_by\_year:2023:2016
#> referenced\_works:https://openalex.org/W4248252387
#> related\_works:https://openalex.org/W188202134
#> is\_paratext:{FALSE}
#> is\_retracted:{FALSE}},
#> 	keywords = {Bibliometrics, Computer science, Data mining, Data science, Database, Programming language, Software, Software engineering, Software tool, Workflow},
#> }
#> 
# }