Make a Zotero friendly format of input data

Make a Zotero friendly format of input data, will add a zotero key.

Usage

ZoteroFormat(
  data = NULL,
  format = NULL,
  prefix = NULL,
  check.structure = FALSE,
  silent = FALSE
)

Arguments

data

A list of metadata or something else, Default: NULL

format

The format of the input data (e.g., 'JSON', 'versions', 'keys). , Default: NULL

prefix

Add a prefix to the metadata (e.g., user/userID), Default: NULL

check.structure

Check that the structure of a data frame is correct according to the Zotero type, Default: FALSE

silent

c2z is noisy, tell it to be quiet, Default: TRUE

Value

A zotero friendly tibble if requested otherwise format the data according to format.

See also

Examples

# \donttest{
  ZoteroFormat(
    list(title = "This is a test", itemType = "document")
  )
#> # A tibble: 1 × 5
#>   key      version title          itemType relations   
#>   <chr>      <dbl> <chr>          <chr>    <list>      
#> 1 WN9F4JGU       0 This is a test document <df [0 × 0]>
# }