Export Zotero items to bibliography

Export Zotero items to a specified format (e.g., BibLaTeX) using a CSL format (e.g., APA7)

Usage

ZoteroExport(
  zotero,
  csl.type = NULL,
  csl.name = "style",
  locale = "en-US",
  format = "biblatex",
  bib.name = "references",
  save.data = FALSE,
  save.path = NULL,
  silent = FALSE
)

Arguments

zotero

A list with information on the specified Zotero library (e.g., id, API key, collections, and items)

csl.type

Specify a CSL type to Official repository for Citation Style Language (CSL), Default: NULL

csl.name

Name of saved CSL file, Default: 'style'

locale

Desired language format of bibliography, Default: 'en-US'

format

Export format of Zotero items, Default: 'biblatex'

bib.name

Name of exported bibliography, Default: 'references'

save.data

Save data (e.g., bibliography) to disk, Default: FALSE

save.path

Location to store data on disk, Default: NULL

silent

c2z is noisy, tell it to be quiet, Default: FALSE

Value

A list with information on the specified Zotero library (e.g., exported items, bibliography and citations)

Details

Please see https://oeysan.github.io/c2z/

See also

Examples

# \donttest{
  # Define Zotero list according to default group
  zotero = Zotero(
    user = FALSE,
    id = "4827927",
    api = "RqlAmlH5l1KPghfCseAq1sQ1",
    item.type = "-attachment || note",
    max.result = 1,
    library = TRUE
  )
#> Searching for collections 
#> Found 2 collections 
#> The Zotero list contains: 2 collections, 0 items, and 0 attachments 
#> Searching for all items in library 
#> Found 1 item 
#> The Zotero list contains: 2 collections, 1 item, and 0 attachments 

  # Export 1 items from the default group
  example <- ZoteroExport(
    zotero
  )
#> Found 1 `biblatex` reference 
  # Display exported
  cat(example$export, fill = 80)
#> 
#> @article{fruhholz_neurocognitive_2021,
#> 	title = {Neurocognitive processing efficiency for discriminating human non-alarm rather than alarm scream calls},
#> 	volume = {19},
#> 	issn = {1545-7885},
#> 	url = {https://dx.plos.org/10.1371/journal.pbio.3000751},
#> 	doi = {10.1371/journal.pbio.3000751},
#> 	abstract = {Across many species, scream calls signal the affective significance of events to other agents. Scream calls were often thought to be of generic alarming and fearful nature, to signal potential threats, with instantaneous, involuntary, and accurate recognition by perceivers. However, scream calls are more diverse in their affective signaling nature than being limited to fearfully alarming a threat, and thus the broader sociobiological relevance of various scream types is unclear. Here we used 4 different psychoacoustic, perceptual decision-making, and neuroimaging experiments in humans to demonstrate the existence of at least 6 psychoacoustically distinctive types of scream calls of both alarming and non-alarming nature, rather than there being only screams caused by fear or aggression. Second, based on perceptual and processing sensitivity measures for decision-making during scream recognition, we found that alarm screams (with some exceptions) were overall discriminated the worst, were responded to the slowest, and were associated with a lower perceptual sensitivity for their recognition compared with non-alarm screams. Third, the neural processing of alarm compared with non-alarm screams during an implicit processing task elicited only minimal neural signal and connectivity in perceivers, contrary to the frequent assumption of a threat processing bias of the primate neural system. These findings show that scream calls are more diverse in their signaling and communicative nature in humans than previously assumed, and, in contrast to a commonly observed threat processing bias in perceptual discriminations and neural processes, we found that especially non-alarm screams, and positive screams in particular, seem to have higher efficiency in speeded discriminations and the implicit neural processing of various scream types in humans.},
#> 	pages = {e3000751},
#> 	number = {4},
#> 	journaltitle = {{PLOS} Biology},
#> 	shortjournal = {{PLoS} Biol},
#> 	author = {Frühholz, Sascha and Dietziker, Joris and Staib, Matthias and Trost, Wiebke},
#> 	editor = {Malmierca, Manuel S.},
#> 	urldate = {2023-10-24},
#> 	date = {2021-04-13},
#> 	langid = {english},
#> }
#> 
# }