Export Zotero items to bibliography

Export Zotero items to a specified format (e.g., BibLaTeX) using a CSL format (e.g., APA7)

Usage

ZoteroExport(
  zotero,
  csl.type = NULL,
  csl.name = "style",
  locale = "en-US",
  format = "biblatex",
  bib.name = "references",
  save.data = FALSE,
  save.path = NULL,
  silent = FALSE
)

Arguments

zotero

A list with information on the specified Zotero library (e.g., id, API key, collections, and items)

csl.type

Specify a CSL type to Official repository for Citation Style Language (CSL), Default: NULL

csl.name

Name of saved CSL file, Default: 'style'

locale

Desired language format of bibliography, Default: 'en-US'

format

Export format of Zotero items, Default: 'biblatex'

bib.name

Name of exported bibliography, Default: 'references'

save.data

Save data (e.g., bibliography) to disk, Default: FALSE

save.path

Location to store data on disk, Default: NULL

silent

c2z is noisy, tell it to be quiet, Default: FALSE

Value

A list with information on the specified Zotero library (e.g., exported items, bibliography and citations)

Details

Please see https://oeysan.github.io/c2z/

See also

Examples

# \donttest{
  # Define Zotero list according to default group
  zotero = Zotero(
    user = FALSE,
    id = "4827927",
    api = "RqlAmlH5l1KPghfCseAq1sQ1",
    item.type = "-attachment || note",
    max.result = 1,
    library = TRUE
  )
#> Searching for collections 
#> Found 11 collections 
#> The Zotero list contains: 11 collections, 0 items, and 0 attachments 
#> Searching for all items in library 
#> Found 6 items 
#> The provided query is limited to 1 result 
#> The Zotero list contains: 11 collections, 1 item, and 0 attachments 

  # Export 1 items from the default group
  example <- ZoteroExport(
    zotero
  )
#> Found 1 `biblatex` reference 
  # Display exported
  cat(example$export, fill = 80)
#> 
#> @article{mwansa_mind_2021,
#> 	title = {Mind the gaps: On the North/South Nexus in the ‘Sport for Development and Peace’ discourse},
#> 	volume = {5},
#> 	issn = {2535-4051},
#> 	url = {https://journals.oslomet.no/index.php/nordiccie/article/view/4464},
#> 	doi = {10.7577/njcie.4464},
#> 	abstract = {In this conceptual article, we present the “Sport for Development and Peace” ({SDP}) discourse as a case of scientific rationalization. First, we shed light on the ongoing theory debate around the “global/local problematique” in globalization and global policy research in comparative and international education. We then link up with the {SDP} discourse and show that academic work mostly features research related to the fact that the majority of the {SDP} programmes and ways of implementing them have been conceptualized in the Global North, yet are to be implemented in the Global South. In that context, we illustrate International Organizations as sites of scientized knowledge production and translation. Scientific rationalization occurs when specialized technical knowledge and management techniques enter the discourse.},
#> 	pages = {23--35},
#> 	number = {3},
#> 	journaltitle = {Nordic Journal of Comparative and International Education ({NJCIE})},
#> 	shortjournal = {Njcie},
#> 	author = {Mwansa, Kabanda and Kiuppis, Florian},
#> 	urldate = {2023-10-05},
#> 	date = {2021-10-26},
#> 	note = {{NVI}: 1
#> Cristin: 2161200},
#> }
#> 
# }